pyproject.toml
[project]
name = "arnold"
version = "0.1.0"
description = "Knowledge-grounded expert exercise system - Cyberdyne Systems Model 101"
authors = [{name = "Brock", email = "user@example.com"}]
requires-python = ">=3.10"
dependencies = [
    "neo4j>=5.14.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pronto>=2.5.4",  # For parsing OBO ontologies
    "anthropic>=0.39.0",  # For Claude API
    "mcp>=1.0.0",  # Model Context Protocol
    "click>=8.1.7",  # For CLI interface
]

[project.scripts]
arnold = "arnold.cli.coach:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
